Add backport of parameters from ansible 2.0
The modules have been developed against devel of ansible, but there are
folks using them with the current 1.9 release. That's fine - they mostly
work. However, there have been a few changes in the openstack
module_utils that seemed fine becasue the modules weren't released in
1.9.

Backport/forwardport these to the modules temporarily with a bunch of
copypasta so that folks using the modules against 1.9 still get working
modules. Once 2.0 is released, remove the function from everything.